backend/requirements.txt
flask
flask-cors
pillow
pymupdf
gunicorn
